ODSC-72811: Enhance Multi-Model Deployment Configuration to Support FT models
